chore: Handle missing data in create_parameter_list function
The code changes in `tasks.py` modify the `create_parameter_list` function to handle cases where there is no data to process. If there is no data and it is not a routine process, a log message is added and an `ENDRUN` signal with a `SUCCESS` message is raised. Otherwise, an error log message is added and a `FAIL` signal is raised.