feat: Update VitaCare DB extraction and loading flow
This commit updates the "sms_dump_vitacare_db" flow in the DataLake - Extração e Carga de Dados pipeline. It includes changes to the tasks for loading local data to the raw folder, transforming the data, and uploading it to the datalake. The flow is now parameterized with the environment and dataset ID. Additionally, a new task for renaming files to a standard format has been added.